+++ Sources/C1X/C1X/SupportFiles/C1X.h
@@ -1,0 +1,7 @@
+#import <Foundation/Foundation.h>
+
+//! Project version number for PianoC1X.
+FOUNDATION_EXPORT double PianoC1XVersionNumber;
+
+//! Project version string for PianoC1X.
+FOUNDATION_EXPORT const unsigned char PianoC1XVersionString[];